Add support for building on production.
Since middleman is actually retarded, we hijack `manipulate_resource_list` to do the actual font compilation; every SASS file _needs_ to have all its dependants generated by the time this step passes, so this is the only viable entry point. Since the font depends on a directory of files, and not any individual file, we must also monkey patch `.template?` and `.binary?` since they only work on regular (non-directory) files. Finally we output the generated fonts as resources in order to keep the Sitemap system happy. `no_hash` is enabled permanently since it is now the job of `:asset_hash`.